requirements.txt
uvicorn==0.17.6
fastapi==0.75.1
sentence-transformers==2.2.1
huggingface_hub==0.16.4
python-dotenv==0.20.0
sqlalchemy==1.4.35
sqlalchemy-utils==0.38.2
psycopg2-binary==2.9.3
pydantic==1.9.0
starlette==0.17.1
pytest==7.1.2
pytest-asyncio==0.18.3